requirements.txt
alabaster==0.7.12
asgiref==3.3.1
astroid==2.5.1
attrs==19.1.0
Babel==2.9.0
backcall==0.1.0
bleach==3.1.0
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.1
cryptography==2.8
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
docutils==0.16
entrypoints==0.3
fastcache==1.1.0
flasgger==0.9.5
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
imagesize==1.2.0
importlib-metadata==3.7.3
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==5.7.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.1
keyring==23.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.3.3
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.4.1
nbformat==4.4.0
nltk==3.4.1
nose==1.3.7
notebook==5.7.8
numpy==1.16.3
numpydoc==1.1.0
packaging==20.9
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
plotly==4.14.3
pprofile==2.0.2
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psutil==5.7.0
py-heat==0.0.6
py-heat-magic==0.0.2
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.0
Pygments==2.3.1
pylint==2.7.2
pyOpenSSL==19.1.0
pyparsing==2.4.0
pyrsistent==0.15.1
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2021.1
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
QtAwesome==1.0.2
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
retrying==1.3.3
rope==0.18.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.11.1
Send2Trash==1.5.0
six==1.12.0
sklearn-contrib-py-earth==0.1.0
snowballstemmer==2.1.0
Sphinx==3.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
sympy==1.4
terminado==0.8.2
testpath==0.4.2
toml==0.10.2
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.4.2
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
xgboost==1.3.3
zipp==3.4.1
